Brings in the 32-bit overflow audit for exascale runs, the VariableKokkos
consolidation, FFT/Install.sh, zero exponents in variable formulas, and
the ghost-cell acquisition change for gridcut > 0.

Three conflicts, all of them the same bugs found independently on both
branches; each resolved by keeping the superset:

  - FixBalanceKokkos::end_of_step() missing CUSTOM_MASK.  This is now the
    third independent discovery of this one (it also came in through
    claude/upstream-issue-667-gwi6sy).  Kept ALL_MASK, which subsumes the
    explicit CELL|CINFO|SINFO|PCELL|CUSTOM list and matches what
    FixAdaptKokkos and FixMoveSurfKokkos already do.

  - FixGridCheckKokkos returning instead of falling through to an
    out-of-bounds d_cells[icell]/d_cinfo[icell] read after flagging an
    invalid cell.  Identical fix, comment-only conflict; kept the merged
    branch's wording, which notes it matches the non-Kokkos FixGridCheck.

  - ParticleKokkos::post_weight() using h_map after k_map.resize() freed
    the buffer it pointed at.  The merged branch pairs the rebind with a
    MAXSMALLINT guard on the per-processor particle count, so kept that
    version.  This branch's other post_weight fixes are unaffected: the
    k_map.modify_host() before the loop, so the mid-loop resize preserves
    the host side, and assigning the clone id to h_map[nlocal] rather
    than h_map[nlocal-1].

UpdateKokkos merged without conflict but not trivially: the merged branch
splits the packed scalar accumulators into an int d_scalars and a bigint
d_scalars_big, and the retry path on this branch resets them between
attempts.  Verified the reset covers both halves, that h_nmigrate stayed
in the int group, and that grow_mlist() still runs before every kernel
launch.

+++ src/KOKKOS/fft2d_kokkos.cpp
@@ -392,8 +392,8 @@ struct fft_plan_2d_kokkos<DeviceType>* FFT2dKokkos<DeviceType>::fft_2d_create_pl
   else {
     first_ilo = 0;
     first_ihi = nfast - 1;
-    first_jlo = me*nslow/nprocs;
-    first_jhi = (me+1)*nslow/nprocs - 1;
+    first_jlo = (long long) me*nslow/nprocs;
+    first_jhi = (long long) (me+1)*nslow/nprocs - 1;
     plan->pre_plan =
       remapKK->remap_2d_create_plan_kokkos(comm,in_ilo,in_ihi,in_jlo,in_jhi,
                   first_ilo,first_ihi,first_jlo,first_jhi,
@@ -424,8 +424,8 @@ struct fft_plan_2d_kokkos<DeviceType>* FFT2dKokkos<DeviceType>::fft_2d_create_pl
     second_jhi = out_jhi;
   }
   else {
-    second_ilo = me*nfast/nprocs;
-    second_ihi = (me+1)*nfast/nprocs - 1;
+    second_ilo = (long long) me*nfast/nprocs;
+    second_ihi = (long long) (me+1)*nfast/nprocs - 1;
     second_jlo = 0;
     second_jhi = nslow - 1;
   }
@@ -640,7 +640,7 @@ struct fft_plan_2d_kokkos<DeviceType>* FFT2dKokkos<DeviceType>::fft_2d_create_pl
     plan->scaled = 0;
   else {
     plan->scaled = 1;
-    plan->norm = 1.0/(nfast*nslow);
+    plan->norm = 1.0/((double) nfast*nslow);
     plan->normnum = (out_ihi-out_ilo+1) * (out_jhi-out_jlo+1);
   }
